+++ packages/next/src/client/on-recoverable-error.ts
@@ -1,10 +1,4 @@
-import isError from '../lib/is-error'
 import { isBailoutToCSRError } from '../shared/lib/lazy-dynamic/bailout-to-csr'
-import { isHydrationFailureError } from './components/is-hydration-error'
-
-function isSkippableHydrationError(error: unknown) {
-  return isError(error) && isHydrationFailureError(error)
-}
 
 export default function onRecoverableError(err: unknown) {
   // Using default react onRecoverableError
@@ -20,7 +14,6 @@ export default function onRecoverableError(err: unknown) {
 
   // Skip certain custom errors which are not expected to be reported on client
   if (isBailoutToCSRError(err)) return
-  if (isSkippableHydrationError(err)) return
 
   defaultOnRecoverableError(err)
 }
