--- test/development/acceptance-app/fixtures/rsc-runtime-errors/app/hydration-mismatch/bad-nesting/page.js
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <p>
-      <p>Nested p tags</p>
-    </p>
-  )
-}
