+++ test/development/acceptance-app/hydration-error.test.ts
@@ -276,4 +276,45 @@ describe('Error overlay for hydration errors', () => {
 
     await cleanup()
   })
+
+  it('should only show one hydration error when bad nesting happened', async () => {
+    const { cleanup, session, browser } = await sandbox(
+      next,
+      new Map([
+        [
+          'app/page.js',
+          outdent`
+            'use client'
+
+            export default function Page() {
+              return (
+                <p>
+                  <p>Nested p tags</p>
+                </p>
+              )
+            }
+          `,
+        ],
+      ])
+    )
+    // const browser = await next.browser('/hydration-mismatch/bad-nesting')
+    // await waitForAndOpenRuntimeError(browser)
+    await session.waitForAndOpenRuntimeError()
+    expect(await session.hasRedbox()).toBe(false)
+
+    const totalErrorCount = await browser
+      .elementByCss('[data-nextjs-dialog-header-total-count]')
+      .text()
+    expect(totalErrorCount).toBe('1')
+
+    const description = await session.getRedboxDescription()
+    expect(description).toContain(
+      'Error: Hydration failed because the initial UI does not match what was rendered on the server.'
+    )
+    expect(description).toContain(
+      'Warning: Expected server HTML to contain a matching <p> in <p>.'
+    )
+
+    await cleanup()
+  })
 })
